[df] Avoid text flooding in test
A test case with a TTree with variable sized arrays was always printing
to screen the entire content of the arrays, flooding the test output
even in the correct case. Since the arrays are quite long this is not
really helpful even in case of failure, we can rely on the output of the
gtest macro.