[df] Lift template requirement for RNTuple Snapshot
A new action helper is introduced to allow calling Snapshot without template
arguments and without the need for JITting to write RNTuple data to disk.

Unlike TTree, RNTuple is actually able to write certain types to disk even
without explicit dictionary information or RTTI. Notably, all the class of
composed types such as `std::vector<std::vector<...>>` and similar, for which
RField is able to create recursively the inner fields if all the components are
known.

This required a slight modification in the logic. A new tag type
`UseNativeDataType` is used in contexts that require `std::type_info` such as
`GetColumnReaders` to signal that we let the RNTuple data source figure out the
type. The tag is also used in the new action helper to retrieve the correct type
name to crete the RField with.