[df] Increase use case coverage of Snapshot without template
* ROOT::Internal::RDF::AddSizeBranches was never considering the possibility of a
column being a leaf coming from an in-memory struct stored via a leaflist (i.e.
`struct A { int a; int b;}` stored via `Branch("br", &br, "a/I:b/I")`). In such
cases, there is no extra size branch to add so the function now skips this type
of column

* The bufSize and splitLevel arguments to various Branch calls were
not considering the related values of the input branch if it exists.

* A specific scenario with a Redefine changing the type of the column to
Snapshot from a C-style array stored with a leaflist to a ROOT::RVec was still
storing the value as C-style array in the output TTree. Now, when a column value
is coming from a define, the CreateCStyleArrayBranch path is never taken, as it
is not allowed to create C-style array columns from defines.

* The case of an input TTree branch containing unsplit objects of polymorphic
types was not considered in SetBranchesHelper and is now added.